New version.

+++ ThotBot/Intent/TrainedIntentions.cs
@@ -2,12 +2,14 @@
 using System.Collections.Generic;
 using System.Text;
 using ThotBot.Learning;
+using ThotBot.Libray.Cognitive;
+using ThotLibrary;
 
 namespace ThotBot.Intent
 {
     public class TrainedIntentions
     {
-        public int Id;
+        public long Id;
         public string Text;
         public string[] TokenizedText;
         public string IntentLabel;
